Implement particle effects

+++ color.h
@@ -4,10 +4,10 @@
 #include <stdint.h>
 
 typedef struct {
-    uint8_t r;
-    uint8_t g;
-    uint8_t b;
-    uint8_t a;
+    double r;
+    double g;
+    double b;
+    double a;
 } Color;
 
 #endif
